requirements.txt
anthropic
httpx
jinja2
pydantic
PyYAML
python-dotenv
playwright
beautifulsoup4
selectolax
tenacity
pytest
